changing Calc.add() from working with ints to working with floats

+++ DemoCalc/src/Calc.java
@@ -1,7 +1,7 @@
 public class Calc {
 
 	/** @return the sum of a and b */
-	public int add(int a, int b) {
+	public float add(float a, float b) {
 		return a+b;
 	}
